model now takes float

+++ raspi/models/models.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 
 class SHT30X(BaseModel):
-    temperature: int
-    humidity: int
+    temperature: float
+    humidity: float
     # ex of an optional attribute
     # age: Optional[int] = None
